OBPIH-5967 Validate that putaway is not completed before removing putawayItems (#4477)

+++ grails-app/controllers/org/pih/warehouse/api/PutawayItemApiController.groovy
@@ -8,6 +8,10 @@
  * You must not remove this notice, or any other, from this software.
  **/
 package org.pih.warehouse.api
+
+import org.pih.warehouse.order.Order
+import org.pih.warehouse.order.OrderItem
+
 /**
  * Should not extend BaseDomainApiController since putawayItem is not a valid domain.
  */
@@ -16,6 +20,16 @@ class PutawayItemApiController {
     def putawayService
 
     def removingItem() {
+        OrderItem orderItem = OrderItem.get(params.id)
+        if (!orderItem) {
+            throw new IllegalArgumentException("No putaway item found with ID ${id}")
+        }
+        Order order = Order.get(orderItem.order.id)
+
+        if (order && Putaway.getPutawayStatus(order.status) == PutawayStatus.COMPLETED) {
+            throw new IllegalArgumentException("Can't remove an item on completed putaway")
+        }
+
         putawayService.deletePutawayItem(params.id)
 
         render status: 204
